add update endpoint for portfolio and programs
